Feat: 게시글 Update,Delete 함수에 현재 member와 게시글 작성자를 비교하는 코드를 추가

+++ src/main/java/com/patientpal/backend/post/service/PostService.java
@@ -50,16 +50,25 @@ public class PostService {
     }
 
     @Transactional
-    public Post updatePost(Long id, PostUpdateRequest updateRequest) {
+    public Post updatePost(Member member, Long id, PostUpdateRequest updateRequest) {
         Post post = postRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException(ErrorCode.POST_NOT_FOUND));
 
+        Member postMember = post.getMember();
+        if (!postMember.equals(member)) {
+            throw new EntityNotFoundException(ErrorCode.AUTHORIZATION_FAILED);
+        }
         post.update(updateRequest.getTitle(), updateRequest.getContent());
-
         return post;
     }
 
-    public void deletePost(Long id) {
+    public void deletePost(Member member, Long id) {
+        Post post = postRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException(ErrorCode.POST_NOT_FOUND));
+        Member postMember = post.getMember();
+        if(!postMember.equals(member)) {
+            throw new EntityNotFoundException(ErrorCode.AUTHORIZATION_FAILED);
+        }
         postRepository.deleteById(id);
     }
